requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
random2==1.0.1
tqdm==4.66.1 